fix: 防止错误的删除临时文件

+++ app.py
@@ -55,13 +55,16 @@ if __name__ == "__main__":
                     args.url, proxy=args.proxy if not args.no_proxy else None, process_num=args.process)
     asyncio.run(down.go())
 
-    subprocess.run(["ffmpeg", "-i", args.url.split("/")
-                    [-1], "-c", "copy", f"{args.name}.mkv"], cwd=os.path.join("video", args.name))
-    # 删除原始文件
-    for filepath in glob.glob(f"video/{args.name}/*.ts"):
-        os.remove(filepath)
-    for filepath in glob.glob(f"video/{args.name}/*.m3u8"):
-        os.remove(filepath)
-    for filepath in glob.glob(f"video/{args.name}/*.json"):
-        os.remove(filepath)
-    logging.info(f'{args.name}\t下载完毕')
+    sp = subprocess.run(["ffmpeg", "-i", args.url.split("/")
+                          [-1], "-c", "copy", f"{args.name}.mkv"], cwd=os.path.join("video", args.name))
+    if sp.returncode != 0:
+        logging.error("合并异常")
+    else:
+        # 删除原始文件
+        for filepath in glob.glob(f"video/{args.name}/*.ts"):
+            os.remove(filepath)
+        for filepath in glob.glob(f"video/{args.name}/*.m3u8"):
+            os.remove(filepath)
+        for filepath in glob.glob(f"video/{args.name}/*.json"):
+            os.remove(filepath)
+        logging.info(f'{args.name}\t下载完毕')
